backend/ai-media-service/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
edge-tts
torch
diffusers
transformers
accelerate
opencv-python-headless
pillow
mutagen
scipy
kafka-python
